requirements.txt
Flask~=3.1.1
logger~=1.4
requests~=2.32.3
pytest~=8.3.5
PyYAML~=6.0.2